update mysql/redis client (from openresty)

+++ lua-libs/redis.lua
@@ -1,11 +1,11 @@
--- Copyright (C) 2012 Yichun Zhang (agentzh)
-
+-- Copyright (C) Yichun Zhang (agentzh), CloudFlare Inc.
+-- oneoo modifed to aLiLua
 
 local sub = string.sub
+local byte = string.byte
 local tcp = cosocket.tcp
 local insert = table.insert
 local concat = table.concat
-local len = string.len
 local null = null
 local pairs = pairs
 local unpack = unpack
@@ -13,37 +13,55 @@ local setmetatable = setmetatable
 local tonumber = tonumber
 local error = error
 
-module(...)
 
-_VERSION = '0.15'
+local ok, new_tab = pcall(require, "table.new")
+if not ok then
+    new_tab = function (narr, nrec) return {} end
+end
+
+
+local _M = new_tab(0, 155)
+_M._VERSION = '0.20'
+
 
 local commands = {
     "append",            "auth",              "bgrewriteaof",
-    "bgsave",            "blpop",             "brpop",
-    "brpoplpush",        "config",            "dbsize",
+    "bgsave",            "bitcount",          "bitop",
+    "blpop",             "brpop",
+    "brpoplpush",        "client",            "config",
+    "dbsize",
     "debug",             "decr",              "decrby",
-    "del",               "discard",           "echo",
+    "del",               "discard",           "dump",
+    "echo",
     "eval",              "exec",              "exists",
     "expire",            "expireat",          "flushall",
     "flushdb",           "get",               "getbit",
     "getrange",          "getset",            "hdel",
     "hexists",           "hget",              "hgetall",
-    "hincrby",           "hkeys",             "hlen",
-    "hmget",             --[[ "hmset", ]]     "hset",
+    "hincrby",           "hincrbyfloat",      "hkeys",
+    "hlen",
+    "hmget",             --[[ "hmset", ]]     "hscan",
+    "hset",
     "hsetnx",            "hvals",             "incr",
-    "incrby",            "info",              "keys",
+    "incrby",            "incrbyfloat",       "info",
+    "keys",
     "lastsave",          "lindex",            "linsert",
     "llen",              "lpop",              "lpush",
     "lpushx",            "lrange",            "lrem",
     "lset",              "ltrim",             "mget",
+    "migrate",
     "monitor",           "move",              "mset",
     "msetnx",            "multi",             "object",
-    "persist",           "ping",              "psubscribe",
-    "publish",           "punsubscribe",      "quit",
+    "persist",           "pexpire",           "pexpireat",
+    "ping",              "psetex",       --[[ "psubscribe", ]]
+    "pttl",
+    "publish",      --[[ "punsubscribe", ]]   "pubsub",
+    "quit",
     "randomkey",         "rename",            "renamenx",
+    "restore",
     "rpop",              "rpoplpush",         "rpush",
     "rpushx",            "sadd",              "save",
-    "scard",             "script",
+    "scan",              "scard",             "script",
     "sdiff",             "sdiffstore",
     "select",            "set",               "setbit",
     "setex",             "setnx",             "setrange",
@@ -51,22 +69,35 @@ local commands = {
     "sismember",         "slaveof",           "slowlog",
     "smembers",          "smove",             "sort",
     "spop",              "srandmember",       "srem",
-    "strlen",            "subscribe",         "sunion",
-    "sunionstore",       "sync",              "ttl",
-    "type",              "unsubscribe",       "unwatch",
+    "sscan",
+    "strlen",       --[[ "subscribe", ]]      "sunion",
+    "sunionstore",       "sync",              "time",
+    "ttl",
+    "type",         --[[ "unsubscribe", ]]    "unwatch",
     "watch",             "zadd",              "zcard",
     "zcount",            "zincrby",           "zinterstore",
     "zrange",            "zrangebyscore",     "zrank",
     "zrem",              "zremrangebyrank",   "zremrangebyscore",
     "zrevrange",         "zrevrangebyscore",  "zrevrank",
+    "zscan",
     "zscore",            "zunionstore",       "evalsha"
+}
+
+
+local sub_commands = {
+    "subscribe", "psubscribe"
+}
+
+
+local unsub_commands = {
+    "unsubscribe", "punsubscribe"
 }
 
 
 local mt = { __index = _M }
 
 
-function new(self)
+function _M.new(self)
     local sock, err = tcp()
     if not sock then
         return nil, err
@@ -75,7 +106,7 @@ function new(self)
 end
 
 
-function set_timeout(self, timeout)
+function _M.set_timeout(self, timeout)
     local sock = self.sock
     if not sock then
         return nil, "not initialized"
@@ -85,29 +116,33 @@ function set_timeout(self, timeout)
 end
 
 
-function connect(self, host, port)
-	self.host = host
-	self.port = port -- oneoo add
+function _M.connect(self, ...)
     local sock = self.sock
     if not sock then
         return nil, "not initialized"
     end
 
-    return sock:connect(host, port)
+    self.subscribed = nil
+
+    return sock:connect(...)
 end
 
 
-function set_keepalive(self, ...)
+function _M.set_keepalive(self, ...)
     local sock = self.sock
     if not sock then
         return nil, "not initialized"
+    end
+
+    if self.subscribed then
+        return nil, "subscribed state"
     end
 
     return sock:setkeepalive(...)
 end
 
 
-function get_reused_times(self)
+function _M.get_reused_times(self)
     local sock = self.sock
     if not sock then
         return nil, "not initialized"
@@ -117,7 +152,7 @@ function get_reused_times(self)
 end
 
 
-function close(self)
+local function close(self)
     local sock = self.sock
     if not sock then
         return nil, "not initialized"
@@ -125,17 +160,21 @@ function close(self)
 
     return sock:close()
 end
+_M.close = close
 
 
-local function _read_reply(sock)
-    local line, err = sock:receive('*l')
+local function _read_reply(self, sock)
+    local line, err = sock:receive()
     if not line then
+        if err == "timeout" and not self.subscribed then
+            sock:close()
+        end
         return nil, err
     end
 
-    local prefix = sub(line, 1, 1)
+    local prefix = byte(line)
 
-    if prefix == "$" then
+    if prefix == 36 then    -- char '$'
         -- print("bulk reply")
 
         local size = tonumber(sub(line, 2))
@@ -145,6 +184,9 @@ local function _read_reply(sock)
 
         local data, err = sock:receive(size)
         if not data then
+            if err == "timeout" then
+                sock:close()
+            end
             return nil, err
         end
 
@@ -155,12 +197,12 @@ local function _read_reply(sock)
 
         return data
 
-    elseif prefix == "+" then
+    elseif prefix == 43 then    -- char '+'
         -- print("status reply")
 
         return sub(line, 2)
 
-    elseif prefix == "*" then
+    elseif prefix == 42 then -- char '*'
         local n = tonumber(sub(line, 2))
 
         -- print("multi-bulk reply: ", n)
@@ -168,27 +210,31 @@ local function _read_reply(sock)
             return null
         end
 
-        local vals = {};
+        local vals = new_tab(n, 0);
+        local nvals = 0
         for i = 1, n do
-            local res, err = _read_reply(sock)
+            local res, err = _read_reply(self, sock)
             if res then
-                insert(vals, res)
+                nvals = nvals + 1
+                vals[nvals] = res
 
             elseif res == nil then
                 return nil, err
 
             else
                 -- be a valid redis error value
-                insert(vals, {false, err})
+                nvals = nvals + 1
+                vals[nvals] = {false, err}
             end
         end
+
         return vals
 
-    elseif prefix == ":" then
+    elseif prefix == 58 then    -- char ':'
         -- print("integer reply")
         return tonumber(sub(line, 2))
 
-    elseif prefix == "-" then
+    elseif prefix == 45 then    -- char '-'
         -- print("error reply: ", n)
 
         return false, sub(line, 2)
@@ -200,25 +246,29 @@ end
 
 
 local function _gen_req(args)
-    local req = {"*", #args, "\r\n"}
+    local nargs = #args
 
-    for i = 1, #args do
+    local req = new_tab(nargs + 1, 0)
+    req[1] = "*" .. nargs .. "\r\n"
+    local nbits = 1
+
+    for i = 1, nargs do
         local arg = args[i]
+        nbits = nbits + 1
 
         if not arg then
-            insert(req, "$-1\r\n")
+            req[nbits] = "$-1\r\n"
 
         else
-            insert(req, "$")
-            insert(req, len(arg))
-            insert(req, "\r\n")
-            insert(req, arg)
-            insert(req, "\r\n")
+            if type(arg) ~= "string" then
+                arg = tostring(arg)
+            end
+            req[nbits] = "$" .. #arg .. "\r\n" .. arg .. "\r\n"
         end
     end
 
     -- it is faster to do string concatenation on the Lua land
-    return req
+    return concat(req)
 end
 
 
@@ -234,34 +284,45 @@ local function _do_cmd(self, ...)
 
     local reqs = self._reqs
     if reqs then
-        insert(reqs, req)
+        reqs[#reqs + 1] = req
         return
     end
 
-    -- print("request: ", table.concat(req, ""))
+    -- print("request: ", table.concat(req))
 
     local bytes, err = sock:send(req)
-	if bytes == nil and self.host then
-		local r = connect(self, self.host, self.port)
-		if r then
-			bytes, err = sock:send(req)
-		end
-	end
     if not bytes then
         return nil, err
     end
 
-    return _read_reply(sock)
+    return _read_reply(self, sock)
 end
 
 
-function read_reply(self)
+local function _check_subscribed(self, res)
+    if type(res) == "table"
+       and (res[1] == "unsubscribe" or res[1] == "punsubscribe")
+       and res[3] == 0
+   then
+        self.subscribed = nil
+    end
+end
+
+
+function _M.read_reply(self)
     local sock = self.sock
     if not sock then
         return nil, "not initialized"
     end
 
-    return _read_reply(sock)
+    if not self.subscribed then
+        return nil, "not subscribed"
+    end
+
+    local res, err = _read_reply(self, sock)
+    _check_subscribed(self, res)
+
+    return res, err
 end
 
 
@@ -275,14 +336,46 @@ for i = 1, #commands do
 end
 
 
-function hmset(self, hashname, ...)
+for i = 1, #sub_commands do
+    local cmd = sub_commands[i]
+
+    _M[cmd] =
+        function (self, ...)
+            self.subscribed = true
+            return _do_cmd(self, cmd, ...)
+        end
+end
+
+
+for i = 1, #unsub_commands do
+    local cmd = unsub_commands[i]
+
+    _M[cmd] =
+        function (self, ...)
+            local res, err = _do_cmd(self, cmd, ...)
+            _check_subscribed(self, res)
+            return res, err
+        end
+end
+
+
+function _M.hmset(self, hashname, ...)
     local args = {...}
     if #args == 1 then
         local t = args[1]
-        local array = {}
+
+        local n = 0
         for k, v in pairs(t) do
-            insert(array, k)
-            insert(array, v)
+            n = n + 2
+        end
+
+        local array = new_tab(n, 0)
+
+        local i = 0
+        for k, v in pairs(t) do
+            array[i + 1] = k
+            array[i + 2] = v
+            i = i + 2
         end
         -- print("key", hashname)
         return _do_cmd(self, "hmset", hashname, unpack(array))
@@ -293,17 +386,17 @@ function hmset(self, hashname, ...)
 end
 
 
-function init_pipeline(self)
-    self._reqs = {}
+function _M.init_pipeline(self, n)
+    self._reqs = new_tab(n or 4, 0)
 end
 
 
-function cancel_pipeline(self)
+function _M.cancel_pipeline(self)
     self._reqs = nil
 end
 
 
-function commit_pipeline(self)
+function _M.commit_pipeline(self)
     local reqs = self._reqs
     if not reqs then
         return nil, "no pipeline"
@@ -317,28 +410,29 @@ function commit_pipeline(self)
     end
 
     local bytes, err = sock:send(reqs)
-	if bytes == nil and self.host then
-		local r = connect(self, self.host, self.port)
-		if r then
-			bytes, err = sock:send(req)
-		end
-	end
     if not bytes then
         return nil, err
     end
 
-    local vals = {}
-    for i = 1, #reqs do
-        local res, err = _read_reply(sock)
+    local nvals = 0
+    local nreqs = #reqs
+    local vals = new_tab(nreqs, 0)
+    for i = 1, nreqs do
+        local res, err = _read_reply(self, sock)
         if res then
-            insert(vals, res)
+            nvals = nvals + 1
+            vals[nvals] = res
 
         elseif res == nil then
+            if err == "timeout" then
+                close(self)
+            end
             return nil, err
 
         else
             -- be a valid redis error value
-            insert(vals, {false, err})
+            nvals = nvals + 1
+            vals[nvals] = {false, err}
         end
     end
 
@@ -346,27 +440,19 @@ function commit_pipeline(self)
 end
 
 
-function array_to_hash(self, t)
-    local h = {}
-    for i = 1, #t, 2 do
+function _M.array_to_hash(self, t)
+    local n = #t
+    -- print("n = ", n)
+    local h = new_tab(0, n / 2)
+    for i = 1, n, 2 do
         h[t[i]] = t[i + 1]
     end
     return h
 end
 
 
-local class_mt = {
-    -- to prevent use of casual module global variables
-    __newindex = function (table, key, val)
-        error('attempt to write to undeclared variable "' .. key .. '"')
-    end
-}
-
-
-function add_commands(...)
+function _M.add_commands(...)
     local cmds = {...}
-    local newindex = class_mt.__newindex
-    class_mt.__newindex = nil
     for i = 1, #cmds do
         local cmd = cmds[i]
         _M[cmd] =
@@ -374,8 +460,7 @@ function add_commands(...)
                 return _do_cmd(self, cmd, ...)
             end
     end
-    class_mt.__newindex = newindex
 end
 
 
-setmetatable(_M, class_mt)
+return _M
